refactor and add more log in route
